Np fp32

+++ src/models/tsp_model/models.py
@@ -52,7 +52,7 @@ class TSPModel(nn.Module):
 
         B, T = xy.size(0), 1
 
-        mask = torch.zeros_like(available).type(torch.float32)
+        mask = torch.zeros_like(available).type(torch.float16)
         mask[available == False] = float('-inf')
 
         if self.encoding is None:
